Send SYSINFO_RET_FAIL on task host.get failed
Currently as in #254 described if vcenter or host is not available the returned zabbix result is still OK instead of FAIL.

It should be FAIL as then we can define an nodata() trigger on an host status or directly vcenter status and make other triggers dependent. A little bit like with Zabbix Agent Status.

I'm not into C so if might done it false? It compiled and seamed to work for me.